realocando as funcoes da main para a controller
