+++ app/src/main/java/com/kleberson/listadecursos/view/MainActivity.kt
@@ -1,14 +1,17 @@
 package com.kleberson.listadecursos.view
 
+import android.content.Context
 import android.os.Bundle
 import android.view.View
 import android.widget.EditText
+import android.widget.TextView
+import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.kleberson.listadecursos.R
-import com.kleberson.listadecursos.controller.PersonController
+import com.kleberson.listadecursos.model.Person
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -22,28 +25,58 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    fun salvar(view: View) {
+    fun salvar(view: View){
         val firstNameInput = findViewById<EditText>(R.id.firstNameInput)
         val lastNameInput = findViewById<EditText>(R.id.lastNameInput)
         val nameCourseInput = findViewById<EditText>(R.id.nameCourseInput)
         val contactInput = findViewById<EditText>(R.id.contactInput)
+        val person = Person(firstNameInput.text.toString(),
+            lastNameInput.text.toString(),
+            nameCourseInput.text.toString(),
+            contactInput.text.toString())
+        savePreferences(person)
 
-        val personController = PersonController(this)
-        personController.salvar(view, firstNameInput, lastNameInput, nameCourseInput, contactInput)
+        firstNameInput.setText("")
+        lastNameInput.setText("")
+        nameCourseInput.setText("")
+        contactInput.setText("")
+
+        Toast.makeText(this, "Dados Salvos", Toast.LENGTH_SHORT).show()
     }
 
-    fun limpar(view: View) {
+    fun limpar(view: View){
         val firstNameInput = findViewById<EditText>(R.id.firstNameInput)
         val lastNameInput = findViewById<EditText>(R.id.lastNameInput)
         val nameCourseInput = findViewById<EditText>(R.id.nameCourseInput)
         val contactInput = findViewById<EditText>(R.id.contactInput)
 
-        val personController = PersonController(this)
-        personController.limpar(view, firstNameInput, lastNameInput, nameCourseInput, contactInput)
+        firstNameInput.setText("")
+        lastNameInput.setText("")
+        nameCourseInput.setText("")
+        contactInput.setText("")
+        clearPreferences()
     }
 
     fun finalizar(view: View) {
-        val personController = PersonController(this)
-        personController.finalizar(view)
+        Toast.makeText(this, "Volte Sempre", Toast.LENGTH_SHORT).show()
+        finish()
     }
+
+    fun savePreferences(person: Person) {
+        val sharedPreferences = getSharedPreferences("user_data", Context.MODE_PRIVATE)
+        val editor = sharedPreferences.edit()
+        editor.putString("firstName", person.firstName.toString())
+        editor.putString("lastName", person.lastName.toString())
+        editor.putString("nameCourse", person.course.toString())
+        editor.putString("contact", person.contact.toString())
+        editor.apply()
+    }
+
+    fun clearPreferences() {
+        val sharedPreferences = getSharedPreferences("user_data", Context.MODE_PRIVATE)
+        val editor = sharedPreferences.edit()
+        editor.clear()
+        editor.apply()
+    }
+
 }
